pyproject.toml
[tool.poetry]
name = "ocr"
version = "0.1.0"
description = ""
authors = ["Marco <marco@mdigennaro.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
pytesseract = "^0.3.10"
pymupdf = "^1.24.9"
pdf2image = "^1.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
